refactor: implemented handlers and modules
