+++ src/modules/timer.cpp
@@ -1,4 +1,4 @@
-#include "../include/timer.hpp"
+#include "../../include/modules/timer.hpp"
 
 void Timer::startTimer() {
   start = std::chrono::high_resolution_clock::now();
